Bugfix (a bug in the original phyton code): We need to shrink the cache in any case, not just in case of a miss.
We need to react to the controller, otherwise the controller can't control the cache and the controller get's crazy.

I try to explain it what happend sometimes:
In edge cases: The controller asks to shrink the cache and starts to send smaller and smaller cache sizes and eventual negative cache sizes.
But we inverted negative cache size's... not good. The cache grows and grows...